Improve Java memory settings recommendations
Recommend an increase in heap size only when both the project and
machine support it for all machine types. Before, this was done only
for Windows. Additionally, recommend larger heaps and set the default
to 2GB, as 8GB of RAM is now the listed requirement on https://developer.android.com/studio to run Android Studio.

Bug: 274154546
Change-Id: I0d68546b1f6ad1c1005d07ff123d0483e0f74680