Remove indefinite delays in compose-designer tests
We've got a few instances of the `while (condition) delay(x)` pattern in
compose-designer tests that are not bounded by a timeout. These tests
might be the source of a target timeout if the condition is never met,
so we need to make sure the timeout is set for each test, so we don't
hit the target timeout.

Bug: 161091273
Test: compose-designer tests
Change-Id: I5a689bb2b25ef6d83f135247261133fa5fceb1f5