[Vitals] Refactor code around the insights annotator
Basically, an "InsightsProviderKey" field is introduced in
"AppInsight" to denote the source of insights, e.g. Crashlytics,
Play Vitals, etc. Then we can simplify the structure of the
insights we pass around. And we won't group insights by its
provider until the we render items for the popup of the gutter
icon action.

Bug: n/a
Test: existing
Change-Id: Ibb95abf344ef4fc103f83c077fa7ec84ab4fde10