Replace usage of DeployTasks on ApplicationDeployer
Current implementation of ApplicationDeployer still invokes
DeployerTasks internally, but this will be refactored in the future.
Using the abstraction provided by ApplicationDeployer allows us to
handle deployment at a higher level for various configurations,
supporting both ASwB and Gradle. Additionally, working with
ApplicationDeployer enables immediate handling of deployment results,
such as the needRestart option, instead of relying on something like
KillAndRestartAppLaunchTask, which was created solely to propagate
results in the old infrastructure.

Furthermore, having an interface like ApplicationDeployer helps in
testing the execution pipeline(see change in tests), as it was
impossible to mock DeployTask due to hardcoded instances.

Fixes: 279870080
Test: LaunchTaskRunnerTest
Change-Id: I5a393acc0d261724e286fc784caded47ba9387b4