Attempt to standardize Connections class
The VariantConnection and Connection classes are merged into a single
Connection interface, retaining most of the fields necessary to the
system and common to both Crashlytics and Vitals.

In addition:
AppInsightsModuleSelector is moved back to firebase. A new one will be
created for Vitals.

VitalsConfigurationManager now tries to match every incoming connection
with a Module. This is done when any of the module's appIds matches the
connection's appId. Its configuration flow is also now dependent on the
login state and the gradle sync listener so it won't try to get android
facets before the project is ready.

Test: new VitalsConfigurationManager test and existing.
Bug: 265153914
Change-Id: Id8060230b26e1ea149c15469dc7f09e94becaa8a