Fix UpdateTest timeout on Windows
A recent platform merge seems to have caused notification.getListener()
to return null, but only on Windows. When that's the case, we now invoke
the hyperlink within a notification using its underlying action.

Unfortunately, this new invocation method doesn't seem to work when the
listener is non-null, so there are now two different ways to invoke a
hyperlink.

Fixes: 280364094
Test: UpdateTest
Change-Id: I20dabe6c617b088a3ce1db760ca528f4b1cec2d3