Avoid potential deadlock in ViewLoader
ViewLoader.createViewFromSuperclass needs a read action to query PSI
information. But we should limit that read action to PSI operations, and
not include the call to createNewInstance. That call does not need a
read lock, and can potentially be blocked waiting for other locks which
could result in a deadlock.

Bug: 265962212
Test: N/A
Change-Id: If8799554ec18022e61af2bdf1473a8a8b0e9022d