Warm up VisualLintService in tests to avoid flakiness
VisualLintService is created the first time it is called. In normal use,
that will be in the error queue of the DesignSurface. But in tests, this
can be triggered while the project is being disposed, which can cause an
exception in the VisualLintService constructor and make the test fail.
To avoid this issue, this calls on VisualLintService in the setup of
affected tests so that if it is called on during dispose, nothing bad
would happen.

Bug: 279929471
Test: test code
Change-Id: I6f0ad3a00e7eb7118dfb12e5110a4997e9ffdb24