Do not run Visual Lint analysis after timeout
When a Visual Lint run times out, the analyzers that hadn't yet run were
left to run, and potentially report their errors as part of the next
Visual Lint run if one was triggered immediately after the one that
timed out. This ensures that the analyzers are not run if the action
that scheduled them has timed out, while still leaving all the necessary
clean up happen.

Bug: 280045813
Test: It should avoid some memory fluctuations of LayoutEditorPreviewTest
Change-Id: I98d705b7f719833c32969e84c9917836b739128c