Avoid JarManager prefetching large JARs repeatedly
This fixes a bug where if a JAR file is larger than the maximum cache
size and prefetching is enabled, it will try to prefetch the full
contents over and over again on every access to the JAR. This slows down
significantly the loading of R classes in large projects where the R.jar
file exceeds the size of the cache.

This change ensures that after the first prefetch load has failed, the
file is tried to be loaded again into the cache. The class file inside
the JAR will still be loaded but prefetch will not be done.

The change also adds a Perfgate test to track the performance of
JarManagerTest moving forward.

Test: JarManagerTest, ProjectSystemClassLoaderTest
Bug: 278783531
Change-Id: Icd7cde9c02de98532fb910fa569aba990e1e341d