Address flakiness in AndroidSafeDeleteTest
This is a speculative fix for test flakiness, because I've not been able
to reproduce the failure locally.

My working theory is that the manifest file has been created but not yet
processed in test runs, leading to safe delete not finding any usages
yet since the XML file isn't checked for references. I'm hoping that by
using the PsiManager to find the PsiFile representing the manifest, we
can force the project to be up-to-date before we run safe delete.

This is in fact what happens in the following test as well (where the
PsiFile is actually needed), which may explain why that test doesn't
seem to be flaking. (Or it's possible we just aren't seeing the flakes
because the test target is noisy, but at least so far I haven't seen any
there.)

Since this is speculative, I'll keep the tracking bug alive for a few
days and watch results, to see if the flake recurs in official builds.

Test: n/a (test code)
Bug: 281709998
Change-Id: I3c9dbe8b99795db72e655090b9253046c459c27a