Register a dedicated gradleJvmResolver for STUDIO_GRADLE_JDK variable
When the STUDIO_GRADLE_JDK environment variable is present this is
automatically used to override the project gradle root JDK config,
only in the case the specified path is valid.

This override happens on AndroidStudioGradleInstallationManager
that end up configuring the java home for the Gradle daemon execution.
However, this is too late since there exist some early validations on
the platform about the specified gradleJvm on LocalGradleExecutionAware
that may cause the Gradle sync to fail even when a valid
STUDIO_GRADLE_JDK path was provided, on those scenarios when the
existing Gradle JDK configuration is invalid.

To address this issue, was exposed on the platform an extension for
the gradleJvmResolver allowing the add custom resolvers for the
gradleJvm like macros but also special cases like our STUDIO_GRADLE_JDK
variable.

<gradleJvmResolver implementation="..StudioJdkResolver" order="first"/>

Bug: 280414674
Test: SingleGradleRootSyncUseStudioGradleJdkIntegrationTest
Change-Id: I8b886297cb4ef00737848c637b091b21d3a52cd9