Make ComposePreviewView responsible for updating previews
The refresh mechanism shouldn't directly ask the surface to update and
refresh the previews. It should go through the ComposePreviewView
instead, so implementors of this interface can add some extra logic if
rendering the previews in components different than the main surface.

This allows, for example, rendering just a single preview when using
Gallery View in Compose Preview Essentials Mode.

Bug: 279393638
Test: compose-designer tests should pass after this refactoring
Change-Id: Ibd224eba02ccd133be6bff196d9110acf8ce7aa8