Add two values to power rail tooltip
This change brings a new feature to the power profiler:
ability to see both the delta and cumulative values for
each timestamp in a track tooltip.

While both data is now able to be viewed in the tooltip,
obviously only one can be displayed. The one displayed
is indicated by the PowerProfilerDisplayMode feature flag
and is referred to as the "primary" value, while the value
used only in the tooltip is referred to as the "secondary"
value. The ability to switch which value (delta or cumulative)
is primary versus secondary gives us flexibility for future
changes.

The formatting of the primary and secondary value is a simple
2x2 table where the track name, value type, and unit serve as
the left column's content (left justified), with the actual
value in the right column (right justified).

Bug: 254116554
Test: added + updated
Change-Id: I31a3f7f42d9905d7fec7a2c2b5051e239ab4b47a