[PSD] Remove AGP initialization code
We no longer recommend using the Project Structure Dialog to upgrade
the AGP version, let alone setting it in a project that does not
already apply the plugin anywhere we can find it.  That makes the
parsedProjectInitializer for the property effectively unreachable
code; remove it (falling through to the default implementation, which
throws if it is ever called).

Bug: 278743759
Test: existing test adjusted
Change-Id: Ia23bf2b7efe592e48b230ee3f55bce148f6148e5