Use project independent way to get android sdk location
We need to provide Android SDK location to Windows Defender check.
The problem with the previous way is that it requires project to be
synced to work. That check happens on project open before project is
synced if it is a new project which results in not providing sdk path.

Android SDK is defined application wide, does not depend on the project,
so we can just extract the location directly instead.

Bug: 280804258
Test: N/a
Change-Id: I6e05b16e21c3a07a1cee4bd0503c586d878b0c7f