Relax casting type for annotation attribute value
In ULC (K1), most of primitive values are modeled as KtLightPsiLiteral,
a subtype of PsiLiteralExpression, which is used to model more complex
form of annotation values.

In SLC (K2), uses of PsiLiteral and PsiLiteralExpression are correctly
separated, and for String value, it's modeled as SymbolPsiLiteral, a
subtype of PsiLiteral.

Relaxing casting type for annotation attribute value from
PsiLiteralExpression to PsiLiteral can handle both LC values.

Bug: 282873126
Test: DataBindingInspectionTest w/o and w/ K2
Change-Id: I15bbe67886b4d04f6e62a0428a779fc22b1cc0ca