Move component renaming tests
Handling all tests using `doRenameComponentTest`. It turns out that the
"move package" tests are no longer applicable, as they test operations
relating to package names on the <manifest> tag. Those are no longer
supported, as the package name is specified in gradle files. As such,
those tests are simply removed.

Bug: 281863312
Test: AndroidComponentRenamingTest
Change-Id: I8d3c45e470e8003f2261405724927eecf7060076