Remove MovePackage* tests from AndroidResourceRenameTest
These tests all create an Activity, reference it from a manifest, and
then move the Activity's containing package, verifying that the manifest
is correctly updated.

I've already ported other tests that verify effectively the same thing,
but by renaming the package rather than moving it. That's just testing
Jetbrains infrastructure effectively, since the Android-specific code
that we run to handle the scenarios is the same. Therefore I'm removing
these tests.

Bug: 281863312
Test: AndroidResourceRenameTest
Change-Id: I3fac63cffa7606062fcb0e190556b4b7602a2394