Removing incorrect assumption from code
The code assumed that the Retry button could not be pressed if another
action was running. Because the methods are not using the right scope
(the suspend coroutineScope) and they are using instead the class level
"coroutineScope", the timeout does not cancel the running job. This
ensures that the behaviour is preserved and that, if the user presses
the button, the previous action is cancelled.

Test: Existing tests pass
Fixes: 203773536
Change-Id: Ic1d00921b8451a70343f8b463c15c841740ee00b