Take device root view size into account when scaling to fit RD rendering
Some devices, in some configurations, only render in part of the screen.
For example a device with a notch in landscape mode won't render over
the notch area.
This needs to be taken into account when translating the rendering from
LI to match the rendering from RD.

This change calculates:
- the left border as the x coordinate of the root view.
- the top border as the y coordinate of the root view.
- the right border as the width of the screen - left border
  + width of root view
- the bottom border as the height of the screen - top border
  + height of root view

Test: LayoutInspectorRendererTest
Bug: 265150325
Change-Id: I052139d4f3b4550d13b704233d7c7febb8f97292