[Ps] Use maven metadata from Maven Central in single artifact search
This involves determining when we are actually semantically performing
a single artifact search, so investigate all uses of SearchQuery to
see what they are doing.  Split the usages into
SingleModuleSearchQuery and ArbitraryModulesSearchQuery, and implement
special behaviour in MavenCentralRepository to handle the SingleModule
case by querying for the existence of maven metadata, falling back to
the existing method if it's not found or completely unparseable.

Fixes: 281973785
Test: added unit tests
Change-Id: I025ab097fe506ec038cf1724cafdfd49842e2cab