[Insights VCS] Show annotations at the updated location
Change-aware annotations are required as we don't
want to show annotations at the old/mismatched
places.

This is achieved by calculating the line shifts for
those untouched lines by diffing the old content and the
current content.

The above work is done in AppInsightsExternalAnnotator#
doAnnotate (which is for long-running activities), we make sure
any read actions on our own if needed.

Bug: 281702889
Test: added
Change-Id: Ibca8a7a974c6af55c5d9ec1869e9704318afaddb