[NPW] Rewrite resolveDependency in terms of Dependency
Most of the uses are to generate a dependency identifier, so make it
explicit that we are working with Dependencies and RichVersions where
appropriate.

In the case of applying a plugin, however, we are not constructing a
Dependency directly (though we are using one using the special Gradle
plugin marker artifact); in that case, use the repositoryUrlManager
directly.

Bug: 279886738
Test: extended existing test
Change-Id: I685910c364811baaf349dc9577b0d42a8bfac7e4