[SDK Index] Change message when multiple issues
... exist for the same library and allow policy issues to be
reported as non-blocking.

When a library has multiple issues we show only one, decided with a
predetermined priority. With this change, we now make it clear that
there are two or more issues instead of picking one.

Also there was an assumption that all policy issues will block
publishing, but that assumption is no longer true; this change uses the
severity from the library for issues that come from the new export.

Test: PolicyIssuesPsdTest, SdkIndexLintTest, SdkIndexPsdTest
Bug: 284994378
Bug: 284992909
Change-Id: I18c561307177599c6407fcca22caac7d9891323d