Sideload tracebox for devices if flag is enabled
Tracebox is sideloaded for devices M,N,O,P if PROFILER_TRACEBOX flag is
enabled. Currently, Perfetto is loaded for devices that are at least P
and this will be modified to load Perfetto for devices at least P only
if flag is disabled.

Also, fixing indentation in StudioFlags.java

Bug: 285412214
Test: added
Change-Id: Icc7737acc20efdd4a6a73eb78a5577902f656a30