Handle rotation from streamed device and emulator separately
The Running Devices Tool Window occasionally fakes the rotation of
streamed devices, which means we can't use the device as source of
truth for the rotation information.

For these devices we need to use the rotation correction provided
by the Running Devices Tool Window.

Unfortunately the rotation correction from Running Devices is
always 0 for emulators.

Therefore when it's 0 we need to calculate our own rotation
correction based on the rotation data provided by the device.

Test: LayoutInspectorRendererTest
Bug: 265150325
Change-Id: Iec9b93b2483760bd6f0148a039885654b1bbe8e4