[BA-Defender] show adjusted notification on project open
We want to have more control over notifications shown about
Windows Defender on project open. Because of that we need to replace
platform's WindowsDefenderCheckerActivity logic with our own.
It was already partly duplicated to make Build Analyzer show this
warning, so this change is only about showing notifications from
our code as well.
Also need to disable WindowsDefenderCheckerActivity in the platform.

Bug: 285867050
Test: Added
Change-Id: Id6a18b15baf7dc8290fac13dbdc54a92e45d75ff