Remove ineffectual template presentation customization
This appears to be a remnant from a long time ago.  The customization
of the Vcs.VcsClone action on the welcome screen (in
setUpWelcomeScreenActions) has been a no-op since around 2019, when
the platform started explicitly setting the presentation text for this
action on the welcome screen.  The customization of the text did have
effect on the underlying action, but that action is not the one that
is displayed in menus either (that is a subclass action, unchanged by
this customization).  So the only effect of this was to change the
name and presentation of the action in the command search.

Also make a method static that could have been but wasn't for no
evident reason.

Bug: n/a
Test: existing
Change-Id: Ibbbe81525cc12853ef43cb56ba02d8fdf192f142