Move more state to flows to make it reactive
This reduces the amount of manual checking and moves it to using flows.
The final goal is to reduce unnecessary refreshes by only looking at the
changes in the declared Previews.

Move filtering in the Preview to use a better defined API that does not
allow for undefined states (like filtering by group and having a single
selection at the same time).

Similarly to above, moving PreviewGroup to have a better definition
where PreviewGroup.Named are the ones defined in the @Preview annotation
and PreviewGroup.all is a separate object to define all the groups. This
ensures better type safety and a cleaner API.

Bug: N/A
Test: //tools/adt/idea/compose-designer/...
Change-Id: I577bf6e823c7b7d4aab195d940dc68596bc9d3d0