[Gradle] Remove GradleCoordinate from GradleDependencyManager
Lint GradleDetector checks and fixes are coupled to IDE Gradle
Dependency management through the policy and mechanism for deciding
where to recommend declarations should be: in build files or in a
catalog.  This coupling means that GradleDependencyManager and its
siblings are sensitive to the change in GradleDetector, so change them
over to use Dependency.  (This means that the conversion between
GradleCoordinate and Dependency now approaches the ModuleSystem
interface layer, which cannot trivially be changed without reworking
those interfaces).

Bug: 257726238
Bug: 259279612
Bug: 242691473
Bug: 279886738
Test: existing tests adjusted
Change-Id: I94ac518148059a9f47c466996fb3c85129a1f2e4