Fix missing main thread name in system trace
The main thread of the selected process' name was displaying
an empty string. Investigation into this showed that this is
because the actual process metadata received from perfetto
contains an empty string for the main thread's name.

To fix this, the main thread's name, if shown to be empty, is
replaced with the process' name.

Fixes: 286011927
Test: added
Change-Id: Iae89b004b7464beb44790e9f080eb9e5a0dad7bf