Rewrite Tests Related to DesugarUtils
There were 4 tests in AndroidPositionManagerTest that targeted the
behavior of DesugarUtils. These tests were poorly designed and
relied on so bad assumptions.

These tests were redesigned and moved to a dedicated test class.

Note that I also changed the sematics of the DesugarUtils functions
to a better pattern. Instead of modifying a mutabale list passed
into the function, it now returns items that should be added by the
caller.

There is still a need to test the integration between
AndroidPositionManager and DesugarUtils. This will be addressed in
another cl.

Bug: NA
Test: Self
Change-Id: I24b55f03690618ebd063c74052ac28104ba5fb20