Update lint baseline
ag/23645766 added support for recognizing usages of annotated
Kotlin objects, e.g. companion objects. This caused some new issues
to be caught, meaning the lint baseline needs to be updated. Very similar issues in the same file/package are already in the baseline,
so it seems reasonable to add these too.

Bug: 282811891
Test: n/a
Change-Id: Ic3a4b6a1b59ea0a3812de6691d4c33795da92112