Increase the idle timeout for benchmarks
I'm seeing daemon restarts even after giving them plenty of idle time. And it's due to the initial runs sometimes surpassing the duration I've found when running with `--config=remote`. Performance characteristics of these tests definitely somehow differ when running on post submit as opposed to running them with `--config=remote`. So my only option (if I can't find the root cause of the differences) is to keep debugging it after submitting.

Bug: 287403162
Change-Id: Ic47ecca7f8ac8af243f19307e24845fc3d883a7e
Test: Test change