Quick fix to stop memory benchmarks crashing
I917c5b98ec729 added a new to the output folder which was less than 3
lines. The current code assumed that all files within the output folder
were heap snapshots and tried to process them.

This change doesn't process files that don't have at least three lines
as these are not heap snapshots. This likely warrents some further
discrimination of files within the folder but we leave this to a future
change.

Bug: N/A
Test: self
Change-Id: I8cc13b8e1ab1e12a17193f6c4223fac62eaae479