Fix incorrect target for Kts/Version Catalog Go To Declaration
The structure of dot-qualified references for Version Catalog entities
is such that getting the grand parent of a literal will always return
a prefix of the whole argument; it is therefore safe to use that
prefix to identify the catalog (and, if necessary, the table within
that catalog).  However, when going to declaration, we should always
consider the whole key, not just a prefix of the fragment that we
happened to have the cursor on.

The behavior differed between Kotlin Script and Groovy Gradle build
files because the IDEA platform also has its own GoToDeclaration
handler for version catalogs, implemented for the special case of a
single Version Catalog in a file named `libs.versions.toml`.  Our
handler is called for non-default names from Groovy build files, and
for all lookups from Kotlin Script build files; make our behavior
consistent with the platform.

Bug: 287643443
Test: added, existing tests adjusted
Change-Id: Ib287ad69b0aa2ab0c2a2cf52c5be7923ee6e178e