Remove obsolete "use filters in TSDKV" flag
Filters in general are now part of the stable assistant, though entirely
in the background to hide steps on startup, so we no longer need the
ability to turn them off.

Specific subfeatures of filters (run on start, WIP filters, redo
filters) remain in development and have their own flags.

Fixes: 288152071
Test: N/A
Change-Id: I94a23ecf2217cf474b99d8c6ca1467e4087051fb