Don't hide status column when grouped
Add visibleWhenGrouped property to Column. If this is set to true for a
column, it will be displayed even when the table is grouped by that
column.

We don't want to hide the status indicator dots when we are grouping by
status. Also, we don't want to indent below the category heading.

Bug: 260752291
Test: Added
Change-Id: Iee075fe03e38e04d1f524064816ffc161d50c128