Show full contents of cell under mouse
Some cells, particularly name and API level (when a codename is
involved), can have contents that do not fit in their allocated width.

One solution is allowing column resizing, which is a bit of a hassle for
the user to manage. Instead, simply expand the cell that is hovered over
to display its entire contents. This is similar to the approach used in
the original Device Manager.

Bug: 287348134
Test: Added
Change-Id: I9e5d1ba14dbc262e3197cb8d4bc8f60483ca28bb