Do not create instance of LayoutInspector in LayoutInspectorManager init method
To keep supporting the standalone Layout Inspector, Layout Inspector is
designed to automatically connect to the available device as soon as it
is created.

This is not desirable in the context of Running Devices, so we should
be careful to create an instance of LayoutInspector only when it's
needed.

Once the standalone Layout Inspector is deprecated we should change
the behaviour of LayoutInspector, to only connect on-demand.

Test: LayoutInspectorManagerTest
Bug: 288635288
Change-Id: I307d2c2886d4d3a11710f639aec8b03b266ef611