Consider data binding when pasting from Resource Manager
When pasting a resource from the Resource Manager in an XML attribute
value, we replace the whole value with the new resource. However, when
the value is using data binding, we should instead keep the data binding
and only replace the content of the data binding. It would still be
possible to replace the entire value by pasting outside of the data
binding brackets.

Fixes: 180693525
Test: tests added
Change-Id: I66e09277f671f941e3ad29895ccf5eb0c329021e